# docs/development-specs/m0-collection-assist/openapi/m0-collection-assist.yaml
openapi: 3.1.0
info:
  title: Blariyo M0 Collection Assist
  version: 0.1.0
  description: Separate opt-in collection routes. Machine routes proxy to /internal/collect; public
    Core routes remain unchanged.
paths:
  /api/v1/admin/collect/sources:
    get:
      operationId: listCollectionSources
      tags:
        - collection-assist
      security:
        - AdminSession: []
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    additionalProperties: false
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/CollectionSource"
                        maxItems: 10000
                        minItems: 0
                    required:
                      - items
                  meta:
                    type: object
                    additionalProperties: false
                    properties:
                      requestId:
                        type: string
                        minLength: 1
                        maxLength: 100
                    required:
                      - requestId
                required:
                  - success
                  - data
                  - meta
        "400":
          $ref: "#/components/responses/CollectionFailure"
        "401":
          $ref: "#/components/responses/CollectionFailure"
        "403":
          $ref: "#/components/responses/CollectionFailure"
        "404":
          $ref: "#/components/responses/CollectionFailure"
        "409":
          $ref: "#/components/responses/CollectionFailure"
        "413":
          $ref: "#/components/responses/CollectionFailure"
        "415":
          $ref: "#/components/responses/CollectionFailure"
        "429":
          $ref: "#/components/responses/CollectionFailure"
        "500":
          $ref: "#/components/responses/CollectionFailure"
        "503":
          $ref: "#/components/responses/CollectionFailure"
  /api/v1/admin/collect/sources/{sourceId}:
    patch:
      operationId: updateCollectionSource
      tags:
        - collection-assist
      security:
        - AdminSession: []
      parameters:
        - in: path
          name: sourceId
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 9007199254740991
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                lockVersion:
                  type: integer
                  minimum: 1
                  maximum: 9007199254740991
                isActive:
                  type: boolean
                fetchMode:
                  type: string
                  enum:
                    - URL_ONLY
                    - LIST_CRAWL
                listUrl:
                  anyOf:
                    - type: string
                      minLength: 1
                      maxLength: 2048
                      format: uri
                      pattern: ^https://
                    - type: "null"
                parserType:
                  type: string
                  enum:
                    - MANUAL
                    - HTML_LIST
                    - RSS
                isListCrawlEnabled:
                  type: boolean
                robotsAllowed:
                  anyOf:
                    - type: boolean
                    - type: "null"
                requestIntervalMs:
                  type: integer
                  minimum: 1000
                  maximum: 2147483647
                dailyFetchLimit:
                  type: integer
                  minimum: 1
                  maximum: 10000
              required:
                - lockVersion
              minProperties: 2
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/CollectionSource"
                  meta:
                    type: object
                    additionalProperties: false
                    properties:
                      requestId:
                        type: string
                        minLength: 1
                        maxLength: 100
                    required:
                      - requestId
                required:
                  - success
                  - data
                  - meta
        "400":
          $ref: "#/components/responses/CollectionFailure"
        "401":
          $ref: "#/components/responses/CollectionFailure"
        "403":
          $ref: "#/components/responses/CollectionFailure"
        "404":
          $ref: "#/components/responses/CollectionFailure"
        "409":
          $ref: "#/components/responses/CollectionFailure"
        "413":
          $ref: "#/components/responses/CollectionFailure"
        "415":
          $ref: "#/components/responses/CollectionFailure"
        "429":
          $ref: "#/components/responses/CollectionFailure"
        "500":
          $ref: "#/components/responses/CollectionFailure"
        "503":
          $ref: "#/components/responses/CollectionFailure"
  /api/v1/admin/collect/candidates:
    get:
      operationId: listCollectionCandidates
      tags:
        - collection-assist
      security:
        - AdminSession: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - PENDING
              - RUNNING
              - NEW
              - FETCH_FAILED
              - APPROVED
              - REJECTED
        - in: query
          name: sourceId
          schema:
            type: integer
            minimum: 1
            maximum: 9007199254740991
        - in: query
          name: discoveryMode
          schema:
            type: string
            enum:
              - MANUAL_URL
        - in: query
          name: duplicateOnly
          schema:
            type: string
            enum:
              - "true"
              - "false"
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            maximum: 10000
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    additionalProperties: false
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/CollectionCandidate"
                        maxItems: 50
                        minItems: 0
                    required:
                      - items
                  meta:
                    type: object
                    additionalProperties: false
                    properties:
                      requestId:
                        type: string
                        minLength: 1
                        maxLength: 100
                      page:
                        type: integer
                        minimum: 1
                        maximum: 10000
                      pageSize:
                        const: 50
                      totalItems:
                        type: integer
                        minimum: 0
                        maximum: 9007199254740991
                      totalPages:
                        type: integer
                        minimum: 0
                        maximum: 9007199254740991
                      hasPrevious:
                        type: boolean
                      hasNext:
                        type: boolean
                    required:
                      - requestId
                      - page
                      - pageSize
                      - totalItems
                      - totalPages
                      - hasPrevious
                      - hasNext
                required:
                  - success
                  - data
                  - meta
        "400":
          $ref: "#/components/responses/CollectionFailure"
        "401":
          $ref: "#/components/responses/CollectionFailure"
        "403":
          $ref: "#/components/responses/CollectionFailure"
        "404":
          $ref: "#/components/responses/CollectionFailure"
        "409":
          $ref: "#/components/responses/CollectionFailure"
        "413":
          $ref: "#/components/responses/CollectionFailure"
        "415":
          $ref: "#/components/responses/CollectionFailure"
        "429":
          $ref: "#/components/responses/CollectionFailure"
        "500":
          $ref: "#/components/responses/CollectionFailure"
        "503":
          $ref: "#/components/responses/CollectionFailure"
    post:
      operationId: createCollectionCandidate
      tags:
        - collection-assist
      security:
        - AdminSession: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9._:-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                originUrl:
                  type: string
                  minLength: 1
                  maxLength: 2048
                  format: uri
                  pattern: ^https://
              required:
                - originUrl
      responses:
        "202":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    additionalProperties: false
                    properties:
                      candidateId:
                        type: integer
                        minimum: 1
                        maximum: 9007199254740991
                      status:
                        type: string
                        enum:
                          - PENDING
                          - RUNNING
                          - NEW
                          - FETCH_FAILED
                          - APPROVED
                          - REJECTED
                      lockVersion:
                        type: integer
                        minimum: 1
                        maximum: 9007199254740991
                      duplicatePostId:
                        anyOf:
                          - type: integer
                            minimum: 1
                            maximum: 9007199254740991
                          - type: "null"
                      reviewedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
                    required:
                      - candidateId
                      - status
                      - lockVersion
                  meta:
                    type: object
                    additionalProperties: false
                    properties:
                      requestId:
                        type: string
                        minLength: 1
                        maxLength: 100
                    required:
                      - requestId
                required:
                  - success
                  - data
                  - meta
        "400":
          $ref: "#/components/responses/CollectionFailure"
        "401":
          $ref: "#/components/responses/CollectionFailure"
        "403":
          $ref: "#/components/responses/CollectionFailure"
        "404":
          $ref: "#/components/responses/CollectionFailure"
        "409":
          $ref: "#/components/responses/CollectionFailure"
        "413":
          $ref: "#/components/responses/CollectionFailure"
        "415":
          $ref: "#/components/responses/CollectionFailure"
        "429":
          $ref: "#/components/responses/CollectionFailure"
        "500":
          $ref: "#/components/responses/CollectionFailure"
        "503":
          $ref: "#/components/responses/CollectionFailure"
  /api/v1/admin/collect/candidates/{candidateId}:
    get:
      operationId: getCollectionCandidate
      tags:
        - collection-assist
      security:
        - AdminSession: []
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 9007199254740991
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    additionalProperties: false
                    properties:
                      candidateId:
                        type: integer
                        minimum: 1
                        maximum: 9007199254740991
                      sourceId:
                        type: integer
                        minimum: 1
                        maximum: 9007199254740991
                      sourceName:
                        type: string
                        minLength: 1
                        maxLength: 50
                      originUrl:
                        type: string
                        minLength: 1
                        maxLength: 2048
                        format: uri
                        pattern: ^https://
                      title:
                        anyOf:
                          - type: string
                            minLength: 1
                            maxLength: 300
                          - type: "null"
                      status:
                        type: string
                        enum:
                          - PENDING
                          - RUNNING
                          - NEW
                          - FETCH_FAILED
                          - APPROVED
                          - REJECTED
                      discoveryMode:
                        type: string
                        enum:
                          - MANUAL_URL
                      imageCandidateCount:
                        type: integer
                        minimum: 0
                        maximum: 20
                      duplicatePostId:
                        anyOf:
                          - type: integer
                            minimum: 1
                            maximum: 9007199254740991
                          - type: "null"
                      postId:
                        anyOf:
                          - type: integer
                            minimum: 1
                            maximum: 9007199254740991
                          - type: "null"
                      rejectReasonCode:
                        anyOf:
                          - type: string
                            minLength: 1
                            maxLength: 100
                            pattern: ^[A-Z][A-Z0-9_]*$
                          - type: "null"
                      fetchErrorCode:
                        anyOf:
                          - type: string
                            minLength: 1
                            maxLength: 100
                            pattern: ^[A-Z][A-Z0-9_]*$
                          - type: "null"
                      requestedAt:
                        type: string
                        format: date-time
                      claimedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
                      fetchedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
                      lockVersion:
                        type: integer
                        minimum: 1
                        maximum: 9007199254740991
                      imageCandidates:
                        type: array
                        items:
                          $ref: "#/components/schemas/CollectionCandidateImage"
                        maxItems: 20
                        minItems: 0
                      warnings:
                        type: array
                        items:
                          type: string
                          minLength: 1
                          maxLength: 100
                          pattern: ^[A-Z][A-Z0-9_]*$
                        maxItems: 20
                        minItems: 0
                    required:
                      - candidateId
                      - sourceId
                      - sourceName
                      - originUrl
                      - title
                      - status
                      - discoveryMode
                      - imageCandidateCount
                      - duplicatePostId
                      - postId
                      - rejectReasonCode
                      - fetchErrorCode
                      - requestedAt
                      - claimedAt
                      - fetchedAt
                      - lockVersion
                      - imageCandidates
                      - warnings
                  meta:
                    type: object
                    additionalProperties: false
                    properties:
                      requestId:
                        type: string
                        minLength: 1
                        maxLength: 100
                    required:
                      - requestId
                required:
                  - success
                  - data
                  - meta
        "400":
          $ref: "#/components/responses/CollectionFailure"
        "401":
          $ref: "#/components/responses/CollectionFailure"
        "403":
          $ref: "#/components/responses/CollectionFailure"
        "404":
          $ref: "#/components/responses/CollectionFailure"
        "409":
          $ref: "#/components/responses/CollectionFailure"
        "413":
          $ref: "#/components/responses/CollectionFailure"
        "415":
          $ref: "#/components/responses/CollectionFailure"
        "429":
          $ref: "#/components/responses/CollectionFailure"
        "500":
          $ref: "#/components/responses/CollectionFailure"
        "503":
          $ref: "#/components/responses/CollectionFailure"
  /api/v1/admin/collect/candidates/{candidateId}/images/{candidateImageId}/preview:
    get:
      operationId: previewCollectionImage
      tags:
        - collection-assist
      security:
        - AdminSession: []
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 9007199254740991
        - in: path
          name: candidateImageId
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 9007199254740991
      responses:
        "200":
          description: Authenticated temporary image
          content:
            image/png:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/CollectionFailure"
        "401":
          $ref: "#/components/responses/CollectionFailure"
        "403":
          $ref: "#/components/responses/CollectionFailure"
        "404":
          $ref: "#/components/responses/CollectionFailure"
        "409":
          $ref: "#/components/responses/CollectionFailure"
        "413":
          $ref: "#/components/responses/CollectionFailure"
        "415":
          $ref: "#/components/responses/CollectionFailure"
        "429":
          $ref: "#/components/responses/CollectionFailure"
        "500":
          $ref: "#/components/responses/CollectionFailure"
        "503":
          $ref: "#/components/responses/CollectionFailure"
  /api/v1/admin/collect/candidates/{candidateId}/retry:
    post:
      operationId: retryCollectionCandidate
      tags:
        - collection-assist
      security:
        - AdminSession: []
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 9007199254740991
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9._:-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                lockVersion:
                  type: integer
                  minimum: 1
                  maximum: 9007199254740991
              required:
                - lockVersion
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    additionalProperties: false
                    properties:
                      candidateId:
                        type: integer
                        minimum: 1
                        maximum: 9007199254740991
                      status:
                        type: string
                        enum:
                          - PENDING
                          - RUNNING
                          - NEW
                          - FETCH_FAILED
                          - APPROVED
                          - REJECTED
                      lockVersion:
                        type: integer
                        minimum: 1
                        maximum: 9007199254740991
                      duplicatePostId:
                        anyOf:
                          - type: integer
                            minimum: 1
                            maximum: 9007199254740991
                          - type: "null"
                      reviewedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
                    required:
                      - candidateId
                      - status
                      - lockVersion
                  meta:
                    type: object
                    additionalProperties: false
                    properties:
                      requestId:
                        type: string
                        minLength: 1
                        maxLength: 100
                    required:
                      - requestId
                required:
                  - success
                  - data
                  - meta
        "400":
          $ref: "#/components/responses/CollectionFailure"
        "401":
          $ref: "#/components/responses/CollectionFailure"
        "403":
          $ref: "#/components/responses/CollectionFailure"
        "404":
          $ref: "#/components/responses/CollectionFailure"
        "409":
          $ref: "#/components/responses/CollectionFailure"
        "413":
          $ref: "#/components/responses/CollectionFailure"
        "415":
          $ref: "#/components/responses/CollectionFailure"
        "429":
          $ref: "#/components/responses/CollectionFailure"
        "500":
          $ref: "#/components/responses/CollectionFailure"
        "503":
          $ref: "#/components/responses/CollectionFailure"
...
components:
  securitySchemes:
    CollectorToken:
      type: http
      scheme: bearer
    AdminSession:
      type: apiKey
      in: cookie
      name: BLARIYO_ADMIN_SESSION
  schemas:
    CollectionSource:
      type: object
      additionalProperties: false
      properties:
        sourceId:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        lockVersion:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        name:
          type: string
          minLength: 1
          maxLength: 50
        baseUrl:
          type: string
          minLength: 1
          maxLength: 2048
          format: uri
          pattern: ^https://
        host:
          type: string
          minLength: 1
          maxLength: 255
        fetchMode:
          type: string
          enum:
            - URL_ONLY
        parserType:
          type: string
          enum:
            - MANUAL
        listUrl:
          type: "null"
        isActive:
          type: boolean
        isListCrawlEnabled:
          const: false
        robotsAllowed:
          anyOf:
            - type: boolean
            - type: "null"
        robotsCheckedAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        requestIntervalMs:
          type: integer
          minimum: 1000
          maximum: 2147483647
        dailyFetchLimit:
          type: integer
          minimum: 1
          maximum: 10000
        lastFetchedAt:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        lastErrorCode:
          anyOf:
            - type: string
              minLength: 1
              maxLength: 100
              pattern: ^[A-Z][A-Z0-9_]*$
            - type: "null"
        disabledReasonCode:
          anyOf:
            - type: string
              minLength: 1
              maxLength: 100
              pattern: ^[A-Z][A-Z0-9_]*$
            - type: "null"
        updatedAt:
          type: string
          format: date-time
      required:
        - sourceId
        - lockVersion
        - name
        - baseUrl
        - host
        - fetchMode
        - parserType
        - listUrl
        - isActive
        - isListCrawlEnabled
        - robotsAllowed
        - robotsCheckedAt
        - requestIntervalMs
        - dailyFetchLimit
        - lastFetchedAt
        - lastErrorCode
        - disabledReasonCode
        - updatedAt
...